accept scientific notation for number tokens
JXG.Dump.toJessie just uses the built-in `toString` method for numbers,
so sometimes they're rendered in scientific notation, e.g. "1.234e-17".

This changes the grammar to allow an exponent on the end of a number
literal.

Numbers that start with a `.` still don't allow an exponent, because
they look for a word boundary on the end. I'm not sure how best to fix
this.